+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - query() now accepts a bind_params argument for parameter binding (#678 thx @clslgrnc)
 
 ### Changed
+- Add consistency param to InfluxDBClient.write_points (#643 thx @RonRothman)
 - Update test suite to add support for Python 3.7 and InfluxDB v1.6.4 and 1.7.4 (#692 thx @clslgrnc)
 - Update classifiers tuple to list in setup.py (#697 thx @Hanaasagi)
 - Update documentation for empty `delete_series` confusion